backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
supabase>=2.3
python-dotenv>=1.0
google-generativeai>=0.3
Pillow>=10.0
python-magic>=0.4

